pyproject.toml
[project]
name = "proct_olis"
version = "1.0"
description = "Default template for PDM package"
authors = [
    {name = "imen-bendai", email = "imen.bendai@outlook.com"},
]
dependencies = [
    "click",
    "pyyaml",
    "polars",
    "pandas",
    "minio",
    "requests",
    "s3fs",
    "dotenv"
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true
[dependency-groups]
lint = [
    "ruff",
    "flake8",
    "black"
]
 
dev = [
    "pytest",
    "ipykernel"
]

[project.scripts]
transformation = "proct_olis.app:main"

[tool.pdm.build]
includes = []
excludes = ["scripts/"]